# Conflicts:
#	backend/service-common/src/main/kotlin/service/k8/CommonResources.kt
#	frontend-web/webclient/app/Applications/Jobs/View.tsx
#	frontend-web/webclient/app/UCloud/index.ts

+++ backend/alerting-service/src/main/kotlin/alerting/services/NetworkTrafficAlerts.kt
@@ -128,7 +128,7 @@ class NetworkTrafficAlerts(
     )
 
     data class Info(
-        val offset: Int,
+        val offset: Long,
         val file: LogFileInfo
     )
